Task 20

+++ 101-binary_tree_levelorder.c
@@ -56,7 +56,7 @@ void print_level(const binary_tree_t *tree, void (*func)(int), size_t level)
  */
 void binary_tree_levelorder(const binary_tree_t *tree, void (*func)(int))
 {
-	size_t height, i;
+	int height, i;
 
 	if (!tree || !func)
 		return;
